+++ src/Connections/FeedConnection.ts
@@ -1,10 +1,8 @@
-import {Appservice, Intent, StateEvent} from "matrix-bot-sdk";
+import {Intent, StateEvent} from "matrix-bot-sdk";
 import { IConnection, IConnectionState, InstantiateConnectionOpts } from ".";
 import { ApiError, ErrCode } from "../api";
-import { BridgeConfigFeeds } from "../Config/Config";
 import { FeedEntry, FeedError, FeedReader} from "../feeds/FeedReader";
 import { Logger } from "matrix-appservice-bridge";
-import { IBridgeStorageProvider } from "../Stores/StorageProvider";
 import { BaseConnection } from "./BaseConnection";
 import markdown from "markdown-it";
 import { Connection, ProvisionConnectionOpts } from "./IConnection";
@@ -27,6 +25,8 @@ export interface LastResultFail {
 export interface FeedConnectionState extends IConnectionState {
     url:    string;
     label?: string;
+    template?: string;
+    notifyOnFailure?: boolean;
 }
 
 export interface FeedConnectionSecrets {
@@ -37,6 +37,11 @@ export type FeedResponseItem = GetConnectionsResponseItem<FeedConnectionState, F
 
 const MAX_LAST_RESULT_ITEMS = 5;
 const VALIDATION_FETCH_TIMEOUT_MS = 5000;
+const MAX_SUMMARY_LENGTH = 512;
+const MAX_TEMPLATE_LENGTH = 1024;
+
+const DEFAULT_TEMPLATE = "New post in $FEEDNAME";
+const DEFAULT_TEMPLATE_WITH_CONTENT = "New post in $FEEDNAME: $LINK"
 
 const nhm = new NodeHtmlMarkdown(
   /* options (optional) */ {},
@@ -49,12 +54,13 @@ export class FeedConnection extends BaseConnection implements IConnection {
     static readonly CanonicalEventType = "uk.half-shot.matrix-hookshot.feed";
     static readonly EventTypes = [ FeedConnection.CanonicalEventType ];
     static readonly ServiceCategory = "feeds";
+    
 
-    public static createConnectionForState(roomId: string, event: StateEvent<any>, {config, as, intent, storage}: InstantiateConnectionOpts) {
+    public static createConnectionForState(roomId: string, event: StateEvent<any>, {config, intent}: InstantiateConnectionOpts) {
         if (!config.feeds?.enabled) {
             throw Error('RSS/Atom feeds are not configured');
         }
-        return new FeedConnection(roomId, event.stateKey, event.content, config.feeds, as, intent, storage);
+        return new FeedConnection(roomId, event.stateKey, event.content, intent);
     }
 
     static async validateUrl(url: string): Promise<void> {
@@ -71,24 +77,38 @@ export class FeedConnection extends BaseConnection implements IConnection {
         }
     }
 
-    static async provisionConnection(roomId: string, _userId: string, data: Record<string, unknown> = {}, {as, intent, config, storage}: ProvisionConnectionOpts) {
-        if (!config.feeds?.enabled) {
-            throw new ApiError('RSS/Atom feeds are not configured', ErrCode.DisabledFeature);
-        }
-
+    static validateState(data: Record<string, unknown> = {}): FeedConnectionState {
         const url = data.url;
         if (typeof url !== 'string') {
             throw new ApiError('No URL specified', ErrCode.BadValue);
         }
-        await FeedConnection.validateUrl(url);
         if (typeof data.label !== 'undefined' && typeof data.label !== 'string') {
             throw new ApiError('Label must be a string', ErrCode.BadValue);
         }
 
-        const state = { url, label: data.label };
+        if (typeof data.template !== 'undefined') {
+            if (typeof data.template !== 'string') {
+                throw new ApiError('Template must be a string', ErrCode.BadValue);
+            }
+            // Sanity to prevent slowing hookshot down with massive templates.
+            if (data.template.length > MAX_TEMPLATE_LENGTH) {
+                throw new ApiError(`Template should not be longer than ${MAX_TEMPLATE_LENGTH} characters`, ErrCode.BadValue);
+            }
+        }
 
-        const connection = new FeedConnection(roomId, url, state, config.feeds, as, intent, storage);
-        await intent.underlyingClient.sendStateEvent(roomId, FeedConnection.CanonicalEventType, url, state);
+
+        return { url, label: data.label, template: data.template };
+    }
+
+    static async provisionConnection(roomId: string, _userId: string, data: Record<string, unknown> = {}, { intent, config }: ProvisionConnectionOpts) {
+        if (!config.feeds?.enabled) {
+            throw new ApiError('RSS/Atom feeds are not configured', ErrCode.DisabledFeature);
+        }
+
+        const state = this.validateState(data);
+        await FeedConnection.validateUrl(state.url);
+        const connection = new FeedConnection(roomId, state.url, state, intent);
+        await intent.underlyingClient.sendStateEvent(roomId, FeedConnection.CanonicalEventType, state.url, state);
 
         return {
             connection,
@@ -119,6 +139,31 @@ export class FeedConnection extends BaseConnection implements IConnection {
         }
     }
 
+    public templateFeedEntry(template: string, entry: FeedEntry) {
+        return template.replace(/(\$[A-Z]+)/g, (token: string) => {
+            switch(token) {
+                case "$FEEDNAME":
+                    return this.state.label || entry.feed.title || entry.feed.url || "";
+                case "$FEEDURL":
+                    return entry.feed.url || "";
+                case "$FEEDTITLE":
+                    return entry.feed.title || "";
+                case "$TITLE":
+                    return entry.title || "";
+                case "$LINK":
+                    return entry.link ? `[${entry.title ?? entry.link}](${entry.link})` : "";
+                case "$AUTHOR":
+                    return entry.author || "";
+                case "$DATE":
+                    return entry.pubdate || "";
+                case "$SUMMARY":
+                    return entry.summary || "";
+                default:
+                    return token;
+            }
+        });
+    }
+
     private hasError = false;
     private readonly lastResults = new Array<LastResultOk|LastResultFail>();
 
@@ -130,10 +175,7 @@ export class FeedConnection extends BaseConnection implements IConnection {
         roomId: string,
         stateKey: string,
         private state: FeedConnectionState,
-        private readonly config: BridgeConfigFeeds,
-        private readonly as: Appservice,
         private readonly intent: Intent,
-        private readonly storage: IBridgeStorageProvider
     ) {
         super(roomId, stateKey, FeedConnection.CanonicalEventType)
         log.info(`Connection ${this.connectionId} created for ${roomId}, ${JSON.stringify(state)}`);
@@ -145,16 +187,19 @@ export class FeedConnection extends BaseConnection implements IConnection {
 
     public async handleFeedEntry(entry: FeedEntry): Promise<void> {
 
-        let entryDetails;
-        if (entry.title && entry.link) {
-            entryDetails = `[${entry.title}](${entry.link})`;
+        let message;
+        if (this.state.template) {
+            message = this.templateFeedEntry(this.state.template, entry);
+        } else if (entry.title && entry.link) {
+            message = this.templateFeedEntry(DEFAULT_TEMPLATE_WITH_CONTENT, entry);
         } else {
-            entryDetails = entry.title || entry.link;
+            message = this.templateFeedEntry(DEFAULT_TEMPLATE, entry);
         }
 
-        let message = `New post in ${this.state.label || entry.feed.title || entry.feed.url}`;
-        if (entryDetails) {
-            message += `: ${entryDetails}`;
+        // This might be massive and cause us to fail to send the message
+        // so confine to a maximum size.
+        if (entry.summary?.length ?? 0 > MAX_SUMMARY_LENGTH) {
+            entry.summary = entry.summary?.substring(0, MAX_SUMMARY_LENGTH) + "…" ?? null;
         }
 
         // SC: overwrite format if enough infos
@@ -184,10 +229,12 @@ export class FeedConnection extends BaseConnection implements IConnection {
             format: "org.matrix.custom.html",
             formatted_body: md.render(message),
             body: message,
+            external_url: entry.link ?? undefined,
+            "uk.half-shot.matrix-hookshot.feeds.item": entry,
         });
     }
 
-    handleFeedSuccess() {
+    public handleFeedSuccess() {
         this.hasError = false;
         this.lastResults.unshift({
             ok: true,
@@ -208,6 +255,10 @@ export class FeedConnection extends BaseConnection implements IConnection {
             // To avoid short term failures bubbling up, if the error is serious, we still bubble.
             return;
         }
+        if (!this.state.notifyOnFailure) {
+            // User hasn't opted into notifications on failure
+            return;
+        }
         if (!this.hasError) {
             await this.intent.sendEvent(this.roomId, {
                 msgtype: 'm.notice',
@@ -218,6 +269,17 @@ export class FeedConnection extends BaseConnection implements IConnection {
         }
     }
 
+    public async provisionerUpdateConfig(userId: string, config: Record<string, unknown>) {
+        // Apply previous state to the current config, as provisioners might not return "unknown" keys.
+        config = { ...this.state, ...config };
+        const validatedConfig = FeedConnection.validateState(config);
+        if (validatedConfig.url !== this.feedUrl) {
+            throw new ApiError('Cannot alter url of existing feed. Please create a new one.', ErrCode.BadValue);
+        }
+        await this.intent.underlyingClient.sendStateEvent(this.roomId, FeedConnection.CanonicalEventType, this.stateKey, validatedConfig);
+        this.state = validatedConfig;
+    }
+    
     // needed to ensure that the connection is removable
     public async onRemove(): Promise<void> {
         log.info(`Removing connection ${this.connectionId}`);